Moves service worker into it a bootstrap script
